# documentation/api/swagger.yaml
openapi: 3.0.0
info:
  title: PulsePlus Game Platform API
  version: 1.0.0
  description: API documentation for the PulsePlus gaming and engagement platform
servers:
  - url: /api
    description: Main API server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
...
    GameSettings:
      type: object
      properties:
        allow_team_creation:
          type: boolean
        max_team_size:
          type: integer
        enable_leaderboard:
          type: boolean
        enable_achievements:
          type: boolean
        enable_badges:
          type: boolean
        enable_shop:
          type: boolean
        custom_rules:
          type: object
...
    Game:
      type: object
      properties:
        sys_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [STANDARD, TOURNAMENT, CUSTOM]
        status:
          type: string
          enum: [ACTIVE, INACTIVE, COMPLETED]
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        settings:
          $ref: '#/components/schemas/GameSettings'
        creator_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - name
        - type
...
paths:
...
  /games:
    get:
      tags:
        - Game Management
      summary: List all games
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, INACTIVE, COMPLETED]
      responses:
        '200':
          description: List of games
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Game'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Game Management
      summary: Create a new game
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Game'
      responses:
        '201':
          description: Game created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /games/{gameId}:
    get:
      tags:
        - Game Management
      summary: Get game details
      security:
        - BearerAuth: []
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Game details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags:
        - Game Management
      summary: Update game details
      security:
        - BearerAuth: []
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Game'
      responses:
        '200':
          description: Game updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /games/{gameId}/stats:
    get:
      tags:
        - Game Management
      summary: Get game statistics
      security:
        - BearerAuth: []
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Game statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_players:
                    type: integer
                  active_players:
                    type: integer
                  total_achievements_earned:
                    type: integer
                  total_xp_earned:
                    type: integer
                  top_performers:
                    type: array
                    items:
                      type: object
                      properties:
                        user_id:
                          type: string
                          format: uuid
                        username:
                          type: string
                        xp:
                          type: integer
                        achievements:
                          type: integer
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # Social Features
...
  /analytics/reports/games/{gameId}:
    post:
      tags:
        - Analytics
      summary: Generate analytics report
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                report_type:
                  type: string
                  enum: [RETENTION, ECONOMY, PROGRESSION, SOCIAL]
                timeframe:
                  type: string
                  enum: [7d, 30d, 90d, ALL]
                  default: 30d
                format:
                  type: string
                  enum: [PDF, CSV, JSON]
                  default: JSON
      responses:
        '200':
          description: Report generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  report_id:
                    type: string
                  download_url:
                    type: string
                  expires_at:
                    type: string
                    format: date-time
...
  /games/{gameId}/settings:
    get:
      tags:
        - Game Management
      summary: Get game settings
      security:
        - BearerAuth: []
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Game settings retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameSettings'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags:
        - Game Management
      summary: Update game settings
      security:
        - BearerAuth: []
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameSettings'
      responses:
        '200':
          description: Game settings updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameSettings'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # Social Features
...
security:
  - BearerAuth: [] 